backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
python-multipart
python-dotenv
alembic
Mako
MarkupSafe
email-validator
bleach>=6.0.0

# AI Chatbot dependencies
openai>=1.0.0  # OpenRouter uses OpenAI-compatible API
google-genai>=1.0.0  # Legacy support (optional)

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
